update demo from fork branch

+++ library/src/main/java/com/liulishuo/filedownloader/download/DownloadStatusCallback.java
@@ -23,6 +23,7 @@ import android.os.HandlerThread;
 import android.os.Message;
 import android.os.SystemClock;
 
+import com.liulishuo.filedownloader.database.FileDownloadDatabase;
 import com.liulishuo.filedownloader.exception.FileDownloadGiveUpRetryException;
 import com.liulishuo.filedownloader.exception.FileDownloadOutOfSpaceException;
 import com.liulishuo.filedownloader.message.MessageSnapshotFlow;
@@ -30,7 +31,6 @@ import com.liulishuo.filedownloader.message.MessageSnapshotTaker;
 import com.liulishuo.filedownloader.model.FileDownloadModel;
 import com.liulishuo.filedownloader.model.FileDownloadStatus;
 import com.liulishuo.filedownloader.services.FileDownloadBroadcastHandler;
-import com.liulishuo.filedownloader.database.FileDownloadDatabase;
 import com.liulishuo.filedownloader.util.FileDownloadLog;
 import com.liulishuo.filedownloader.util.FileDownloadProperties;
 import com.liulishuo.filedownloader.util.FileDownloadUtils;
@@ -38,6 +38,7 @@ import com.liulishuo.filedownloader.util.FileDownloadUtils;
 import java.io.File;
 import java.io.IOException;
 import java.util.concurrent.TimeUnit;
+import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicLong;
 import java.util.concurrent.locks.LockSupport;
 
@@ -135,7 +136,7 @@ public class DownloadStatusCallback implements Handler.Callback {
 
         callbackMinIntervalBytes = calculateCallbackMinIntervalBytes(totalLength,
                 callbackProgressMaxCount);
-        needSetProcess = true;
+        needSetProcess.compareAndSet(false, true);
     }
 
     void onMultiConnection() {
@@ -147,28 +148,27 @@ public class DownloadStatusCallback implements Handler.Callback {
     private volatile long lastCallbackTimestamp = 0;
 
     private final AtomicLong callbackIncreaseBuffer = new AtomicLong();
+    private final AtomicBoolean needCallbackProgressToUser = new AtomicBoolean(false);
 
     void onProgress(long increaseBytes) {
         callbackIncreaseBuffer.addAndGet(increaseBytes);
         model.increaseSoFar(increaseBytes);
 
-
         final long now = SystemClock.elapsedRealtime();
 
-        final boolean isNeedCallbackToUser = isNeedCallbackToUser(now);
+        inspectNeedCallbackToUser(now);
 
         if (handler == null) {
             // direct
-            handleProgress(now, isNeedCallbackToUser);
-        } else if (isNeedCallbackToUser) {
+            handleProgress();
+        } else if (needCallbackProgressToUser.get()) {
             // flow
             sendMessage(handler.obtainMessage(FileDownloadStatus.progress));
         }
     }
 
-    void onRetry(Exception exception, int remainRetryTimes, long invalidIncreaseBytes) {
+    void onRetry(Exception exception, int remainRetryTimes) {
         this.callbackIncreaseBuffer.set(0);
-        model.increaseSoFar(-invalidIncreaseBytes);
 
         if (handler == null) {
             // direct
@@ -231,7 +231,7 @@ public class DownloadStatusCallback implements Handler.Callback {
         } else if (contentLength == TOTAL_VALUE_IN_CHUNKED_RESOURCE) {
             return CALLBACK_SAFE_MIN_INTERVAL_BYTES;
         } else {
-            final long minIntervalBytes = contentLength / (callbackProgressMaxCount + 1);
+            final long minIntervalBytes = contentLength / (callbackProgressMaxCount);
             return minIntervalBytes <= 0 ? CALLBACK_SAFE_MIN_INTERVAL_BYTES : minIntervalBytes;
         }
     }
@@ -292,6 +292,7 @@ public class DownloadStatusCallback implements Handler.Callback {
         final String targetPath = model.getTargetFilePath();
 
         final File tempFile = new File(tempPath);
+        boolean renameFailed = true;
         try {
             final File targetFile = new File(targetPath);
 
@@ -310,14 +311,15 @@ public class DownloadStatusCallback implements Handler.Callback {
                 }
             }
 
-            if (!tempFile.renameTo(targetFile)) {
+            renameFailed = !tempFile.renameTo(targetFile);
+            if (renameFailed) {
                 throw new IOException(FileDownloadUtils.formatString(
                         "Can't rename the  temp downloaded file(%s) to the target file(%s)",
                         tempPath, targetPath
                 ));
             }
         } finally {
-            if (tempFile.exists()) {
+            if (renameFailed && tempFile.exists()) {
                 if (!tempFile.delete()) {
                     FileDownloadLog.w(this,
                             "delete the temp file(%s) failed, on completed downloading.",
@@ -335,7 +337,7 @@ public class DownloadStatusCallback implements Handler.Callback {
         try {
             switch (status) {
                 case FileDownloadStatus.progress:
-                    handleProgress(SystemClock.elapsedRealtime(), true);
+                    handleProgress();
                     break;
                 case FileDownloadStatus.retry:
                     handleRetry((Exception) msg.obj, msg.arg1);
@@ -352,24 +354,26 @@ public class DownloadStatusCallback implements Handler.Callback {
         return true;
     }
 
-    private volatile boolean needSetProcess;
+    private final AtomicBoolean needSetProcess = new AtomicBoolean(false);
 
-    private void handleProgress(final long now,
-                                final boolean isNeedCallbackToUser) {
+    private void handleProgress() {
         if (model.getSoFar() == model.getTotal()) {
             database.updateProgress(model.getId(), model.getSoFar());
             return;
         }
 
-        if (needSetProcess) {
-            needSetProcess = false;
+        if (needSetProcess.compareAndSet(true, false)) {
+            if (FileDownloadLog.NEED_LOG) {
+                FileDownloadLog.i(this, "handleProgress update model's status with progress");
+            }
             model.setStatus(FileDownloadStatus.progress);
         }
 
-        if (isNeedCallbackToUser) {
-            lastCallbackTimestamp = now;
+        if (needCallbackProgressToUser.compareAndSet(true, false)) {
+            if (FileDownloadLog.NEED_LOG) {
+                FileDownloadLog.i(this, "handleProgress notify user progress status");
+            }
             onStatusChanged(FileDownloadStatus.progress);
-            callbackIncreaseBuffer.set(0);
         }
     }
 
@@ -445,20 +449,25 @@ public class DownloadStatusCallback implements Handler.Callback {
         onStatusChanged(FileDownloadStatus.error);
     }
 
-    private boolean isFirstCallback = true;
+    private final AtomicBoolean isFirstCallback = new AtomicBoolean(true);
 
-    private boolean isNeedCallbackToUser(final long now) {
-        if (isFirstCallback) {
-            isFirstCallback = false;
-            return true;
+    private void inspectNeedCallbackToUser(final long now) {
+        final boolean needCallback;
+        if (isFirstCallback.compareAndSet(true, false)) {
+            needCallback = true;
+        } else {
+            final long callbackTimeDelta = now - lastCallbackTimestamp;
+            needCallback = (callbackMinIntervalBytes != NO_ANY_PROGRESS_CALLBACK
+                    && callbackIncreaseBuffer.get() >= callbackMinIntervalBytes)
+                    && (callbackTimeDelta >= callbackProgressMinInterval);
         }
-
-        final long callbackTimeDelta = now - lastCallbackTimestamp;
-
-
-        return (callbackMinIntervalBytes != NO_ANY_PROGRESS_CALLBACK
-                && callbackIncreaseBuffer.get() >= callbackMinIntervalBytes)
-                && (callbackTimeDelta >= callbackProgressMinInterval);
+        if (needCallback && needCallbackProgressToUser.compareAndSet(false, true)) {
+            if (FileDownloadLog.NEED_LOG) {
+                FileDownloadLog.i(this, "inspectNeedCallbackToUser need callback to user");
+            }
+            lastCallbackTimestamp = now;
+            callbackIncreaseBuffer.set(0);
+        }
     }
 
     private void onStatusChanged(final byte status) {
